Revert "Register waterproof"
This reverts commit 61d1bd005b355354bfef7425547f92410f4f1af9.

+++ VecozoWep/Models/TeamVM.cs
@@ -29,7 +29,7 @@ namespace VecozoWeb.Models
             Id = team.Id;
             Kleur = team.Kleur;
             Taak = team.Taak;
-            GemRating = (float)team.GemRating;
+            GemRating = team.GemRating;
         }
 
         public override string ToString()
